Codes/django/requirements.txt
apispec==0.39.0
argcomplete==3.2.2
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.34.55
botocore==1.34.55
certifi==2024.2.2
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==2.2.4
django-ckeditor==6.3.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==2.2.0
django-js-asset==2.0.0
django-s3-storage==0.13.9
django-ses==3.5.2
django-taggit==2.1.0
django-tinymce==3.6.1
djangorestframework==3.10.2
drf-yasg==1.9.1
durationpy==0.6
filelock==3.13.1
future==1.0.0
hjson==3.1.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonfield==3.1.0
kappa==0.6.0
mailjet-rest==1.3.4
Markdown==3.1.1
MarkupSafe==2.1.5
numpy==1.24.4
openapi-codec==1.3.2
pandas==1.1.1
pep517==0.13.1
pillow==10.2.0
pip-tools==6.3.1
pipenv==2023.12.1
pkg_resources==0.0.0
placebo==0.10.0
platformdirs==4.2.0
psycopg2-binary==2.8.6
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2019.2
PyYAML==5.3.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
scikit-learn==0.23.2
scipy==1.10.1
sentry-sdk==1.40.6
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.3.0
stripe==8.5.0
text-unidecode==1.3
threadpoolctl==3.3.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.2
troposphere==4.6.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.25.1
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.54.0
zappa-django-utils==0.4.1
